requirements.txt
python-telegram-bot==13.15
Telethon==1.24.0
